feat: move sql out
