+++ packages/database/src/api/sql/dump.ts
@@ -1,4 +1,4 @@
 import { config } from '@modtree/typeorm-config'
-import { sql } from '../../sql'
+import { sql } from '@modtree/sql'
 
 sql.dump(config.database)
