+++ packages/database/src/api/sql/reset/all.ts
@@ -1,5 +1,5 @@
 import { config } from '@modtree/typeorm-config'
-import { sql } from '../../../sql'
+import { sql } from '@modtree/sql'
 import { analyze } from '../../analyze'
 
 analyze(() => sql.dropDatabase(config.database))
