+++ packages/database/src/api/sql/reset/degree.ts
@@ -1,4 +1,4 @@
 import { config } from '@modtree/typeorm-config'
-import { sql } from '../../../sql'
+import { sql } from '@modtree/sql'
 
 sql.dropTables(config.database, ['degree_modules_module', 'degree'])
