+++ packages/database/src/api/sql/restore-default.ts
@@ -1,4 +1,4 @@
 import { config } from '@modtree/typeorm-config'
-import { sql } from '../../sql'
+import { sql } from '@modtree/sql'
 
 sql.restoreFromFile(config.database, 'postgres-modules-only.sql')
