+++ packages/database/src/api/sql/restore.ts
@@ -1,4 +1,4 @@
 import { config } from '@modtree/typeorm-config'
-import { sql } from '../../sql'
+import { sql } from '@modtree/sql'
 
 sql.restorePrompted(config.database)
