+++ packages/database/tests/environment.ts
@@ -3,7 +3,7 @@ import { Agent } from 'http'
 import { DeleteResult, DataSource, Repository } from 'typeorm'
 import { ModtreeApiResponse, InitProps, Repositories } from '@modtree/types'
 import { config } from '@modtree/typeorm-config'
-import { sql } from '@sql'
+import { sql } from '@modtree/sql'
 import {
   getModuleRepository,
   getModuleCondensedRepository,
