+++ packages/database/tests/restore-server.ts
@@ -1,4 +1,4 @@
 import { config } from '@modtree/typeorm-config'
-import { sql } from '../src/sql'
+import { sql } from '@modtree/sql'
 
 sql.restoreFromFile(config.database, config.restoreSource)
