+++ packages/sql/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@modtree/sql",
+  "version": "0.0.1",
+  "license": "MIT",
+  "author": "modtree",
+  "description": "modtree SQL utilities",
+  "main": "./dist/index.js",
+  "type": "commonjs",
+  "dependencies": {
+    "@modtree/utils": "*",
+    "@modtree/typeorm-config": "*",
+    "@modtree/types": "*"
+  },
+  "scripts": {
+    "lint": "eslint --fix --ext .ts .",
+    "build": "tsc"
+  }
+}
