+++ packages/types/index.d.ts
@@ -5,3 +5,4 @@ export * as ModtreeApiResponse from './api-response'
 export * from './controller'
 export * from './repository'
 export * from './typeorm'
+export * from './sql'
